+++ src/components/common/CreateUserWizard/StepUserName.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { defineMessages } from 'react-intl';
 import { useNavigate } from 'react-router-dom';
 
-import { HookFormInput as Input, ActionHookForm } from '~shared/Fields';
+import { HookFormInput as Input, ActionForm } from '~shared/Fields';
 import { ActionTypes } from '~redux/index';
 import { WizardStepProps } from '~shared/Wizard';
 import { mergePayload } from '~utils/actions';
@@ -63,7 +63,7 @@ const StepUserName = ({
   const handleSuccess = () => navigate(LANDING_PAGE_ROUTE, { replace: true });
 
   return (
-    <ActionHookForm<UserWizardStep2>
+    <ActionForm<UserWizardStep2>
       actionType={ActionTypes.USERNAME_CREATE}
       onSuccess={handleSuccess}
       transform={transform}
@@ -85,7 +85,7 @@ const StepUserName = ({
           }
         />
       )}
-    </ActionHookForm>
+    </ActionForm>
   );
 };
