+++ src/components/common/Dialogs/EditColonyDetailsDialog/EditColonyDetailsDialog.tsx
@@ -5,7 +5,7 @@ import { defineMessages } from 'react-intl';
 
 import { MAX_ANNOTATION_LENGTH } from '~constants';
 import Dialog, { DialogProps, ActionDialogProps } from '~shared/Dialog';
-import { ActionHookForm as Form } from '~shared/Fields';
+import { ActionForm } from '~shared/Fields';
 
 import { ActionTypes } from '~redux';
 import { WizardDialogType } from '~hooks';
@@ -66,7 +66,7 @@ const EditColonyDetailsDialog = ({
 
   return (
     <Dialog cancel={cancel}>
-      <Form<FormValues>
+      <ActionForm<FormValues>
         defaultValues={{
           forceAction: false,
           colonyDisplayName: metadata?.displayName,
@@ -91,7 +91,7 @@ const EditColonyDetailsDialog = ({
           isForce={isForce}
           setIsForce={setIsForce}
         />
-      </Form>
+      </ActionForm>
     </Dialog>
   );
 };
