+++ src/components/common/Dialogs/EditDomainDialog/EditDomainDialog.tsx
@@ -5,7 +5,7 @@ import { defineMessages } from 'react-intl';
 
 import { MAX_ANNOTATION_LENGTH } from '~constants';
 import Dialog, { DialogProps, ActionDialogProps } from '~shared/Dialog';
-import { ActionHookForm as Form } from '~shared/Fields';
+import { ActionForm } from '~shared/Fields';
 
 import { ActionTypes } from '~redux';
 import { WizardDialogType } from '~hooks';
@@ -90,7 +90,7 @@ const EditDomainDialog = ({
 
   return (
     <Dialog cancel={cancel}>
-      <Form<FormValues>
+      <ActionForm<FormValues>
         defaultValues={{
           forceAction: false,
           domainName: name,
@@ -113,7 +113,7 @@ const EditDomainDialog = ({
           handleIsForceChange={setIsForce}
           isForce={isForce}
         />
-      </Form>
+      </ActionForm>
     </Dialog>
   );
 };
