+++ src/components/common/Dialogs/UnlockTokenDialog/UnlockTokenDialog.tsx
@@ -4,7 +4,7 @@ import { string, object, boolean, InferType } from 'yup';
 
 import { MAX_ANNOTATION_LENGTH } from '~constants';
 import Dialog, { ActionDialogProps, DialogProps } from '~shared/Dialog';
-import { ActionHookForm as Form } from '~shared/Fields';
+import { ActionForm } from '~shared/Fields';
 import { ActionTypes } from '~redux/index';
 import { pipe, withMeta, withKey, mapPayload } from '~utils/actions';
 import { WizardDialogType } from '~hooks';
@@ -53,7 +53,7 @@ const UnlockTokenDialog = ({
 
   return (
     <Dialog cancel={cancel}>
-      <Form<FormValues>
+      <ActionForm<FormValues>
         defaultValues={{
           forceAction: false,
           annotationMessage: '',
@@ -75,7 +75,7 @@ const UnlockTokenDialog = ({
           handleIsForceChange={setIsForce}
           isForce={isForce}
         />
-      </Form>
+      </ActionForm>
     </Dialog>
   );
 };
