+++ src/components/frame/TokenActivation/TokensTab/ChangeTokenStateForm.tsx
@@ -6,10 +6,7 @@ import { number, object, InferType } from 'yup';
 
 import { toFinite } from '~utils/lodash';
 import Button from '~shared/Button';
-import {
-  ActionHookForm as ActionForm,
-  HookFormInput as Input,
-} from '~shared/Fields';
+import { ActionForm, HookFormInput as Input } from '~shared/Fields';
 import Numeral from '~shared/Numeral';
 import { Tooltip } from '~shared/Popover';
 import { ActionTypes } from '~redux';
