+++ src/components/shared/Fields/Form/index.ts
@@ -1,2 +1,2 @@
-export { default as HookForm } from './HookForm';
-export { default as ActionHookForm, OnSuccess } from './ActionHookForm';
+export { default as Form } from './Form';
+export { default as ActionForm, OnSuccess } from './ActionForm';
